memory/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
pytest
ipykernel
mem0ai